refactor(cli): finish serve kebab-case filenames
Rename the remaining non-kebab-case serve files and tests, update current imports and developer docs, and remove migrated serve entries from the legacy filename allowlist.
