+++ docs/design/daemon-acp-http/README.md
@@ -153,13 +153,13 @@ client  ────────────────────────
 
 ### 3.1 New module layout (`packages/cli/src/serve/acpHttp/`)
 
-| File                    | Responsibility                                                                                                                                                                              |
-| ----------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `index.ts`              | `mountAcpHttp(app, bridge, opts)` — registers `/acp` routes on the existing Express app.                                                                                                    |
-| `connectionRegistry.ts` | `Acp-Connection-Id` → `AcpConnection` (connection SSE writer, `Map<sessionId, SessionStream>`, pending agent→client requests by JSON-RPC id, monotonic id allocator). TTL + DELETE cleanup. |
-| `jsonRpc.ts`            | JSON-RPC 2.0 parse/validate/serialize helpers; error codes (`-32600` etc.); `_qwen/` namespace guard.                                                                                       |
-| `dispatch.ts`           | Maps inbound JSON-RPC methods → `HttpAcpBridge` calls. Maps `BridgeEvent`s → outbound JSON-RPC frames. The translation table (§4).                                                          |
-| `sseStream.ts`          | Long-lived SSE writer (reuses the backpressure/heartbeat pattern from `server.ts`). Distinct from REST `/events` (different framing: full JSON-RPC objects, not qwen event envelopes).      |
+| File                     | Responsibility                                                                                                                                                                              |
+| ------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `index.ts`               | `mountAcpHttp(app, bridge, opts)` — registers `/acp` routes on the existing Express app.                                                                                                    |
+| `connection-registry.ts` | `Acp-Connection-Id` → `AcpConnection` (connection SSE writer, `Map<sessionId, SessionStream>`, pending agent→client requests by JSON-RPC id, monotonic id allocator). TTL + DELETE cleanup. |
+| `json-rpc.ts`            | JSON-RPC 2.0 parse/validate/serialize helpers; error codes (`-32600` etc.); `_qwen/` namespace guard.                                                                                       |
+| `dispatch.ts`            | Maps inbound JSON-RPC methods → `HttpAcpBridge` calls. Maps `BridgeEvent`s → outbound JSON-RPC frames. The translation table (§4).                                                          |
+| `sse-stream.ts`          | Long-lived SSE writer (reuses the backpressure/heartbeat pattern from `server.ts`). Distinct from REST `/events` (different framing: full JSON-RPC objects, not qwen event envelopes).      |
 
 No change to `bridge.ts` / `eventBus.ts` (additive consumer only).
 
@@ -326,8 +326,8 @@ thin compat shim over `/acp` (separate, later PR).
 
 ## 10. Implementation & verification log (v1)
 
-Implemented in `packages/cli/src/serve/acpHttp/` (`jsonRpc.ts`, `sseStream.ts`,
-`connectionRegistry.ts`, `dispatch.ts`, `index.ts`), mounted from `server.ts`
+Implemented in `packages/cli/src/serve/acpHttp/` (`json-rpc.ts`, `sse-stream.ts`,
+`connection-registry.ts`, `dispatch.ts`, `index.ts`), mounted from `server.ts`
 via `mountAcpHttp(app, bridge, { boundWorkspace })`.
 
 ### Automated (`packages/cli/src/serve/acpHttp/*.test.ts`)
@@ -496,8 +496,8 @@ Another reviewer pass (qwen3.7-max). Suite **30 tests**, live re-verified.
 | F3  | **P2**   | 503 connection-cap rejection had no stderr log.                                                                                                                                                                | `writeStderrLine` with the cap value.                                                                                                                                                                       |
 | F4  | **P2**   | `_qwen/notify stream_error` spread let `event.data.kind` shadow the discriminator.                                                                                                                             | Spread first, then `kind: 'stream_error'`.                                                                                                                                                                  |
 | F5  | **P2**   | `MAX_WORKSPACE_PATH_LENGTH` redeclared (`= 4096`) vs the canonical `fs/paths.js`.                                                                                                                              | Import from `../fs/paths.js` (no divergence).                                                                                                                                                               |
-| F6  | **P2**   | `isObjectParams` duplicated `jsonRpc.isObject`.                                                                                                                                                                | Import `isObject`.                                                                                                                                                                                          |
-| F7  | **P2**   | Raw `process.stderr.write` in `index.ts`/`sseStream.ts` vs `writeStderrLine` elsewhere.                                                                                                                        | Unified on `writeStderrLine` across the module.                                                                                                                                                             |
+| F6  | **P2**   | `isObjectParams` duplicated `json-rpc.isObject`.                                                                                                                                                               | Import `isObject`.                                                                                                                                                                                          |
+| F7  | **P2**   | Raw `process.stderr.write` in `index.ts`/`sse-stream.ts` vs `writeStderrLine` elsewhere.                                                                                                                       | Unified on `writeStderrLine` across the module.                                                                                                                                                             |
 
 ---
