+++ eslint.legacy-filenames.mjs
@@ -82,17 +82,13 @@ export const legacyFilenames = [
   'cronScheduler',
   'customBanner',
   'customMatchers',
-  'daemonStatusProvider',
-  'DaemonTuiAdapter',
   'dangerousRules',
   'DataProcessor',
   'debugLogger',
-  'debugMode',
   'deepMerge',
   'deleteCommand',
   'denialTracking',
   'detectTodoChanges',
-  'deviceFlow',
   'dialogScopeUtils',
   'diffCommand',
   'diffOptions',
@@ -109,7 +105,6 @@ export const legacyFilenames = [
   'editorSettingsManager',
   'envInterpolator',
   'environmentContext',
-  'envSnapshot',
   'envVarResolver',
   'errorCodes',
   'errorHandler',
@@ -205,7 +200,6 @@ export const legacyFilenames = [
   'LlmRewriter',
   'loadedSettingsAdapter',
   'loggingContentGenerator',
-  'loopbackBinds',
   'loopDetectionService',
   'lowlightLoader',
   'LruCache',
@@ -283,7 +277,6 @@ export const legacyFilenames = [
   'quitCommand',
   'quotaErrorDetection',
   'qwenContentGenerator',
-  'qwenDeviceFlowProvider',
   'qwenIgnoreParser',
   'qwenOAuth2',
   'rateLimit',
@@ -309,7 +302,6 @@ export const legacyFilenames = [
   'ripgrepUtils',
   'rulesDiscovery',
   'runBudget',
-  'runQwenServe',
   'runtimeDiagnostics',
   'runtimeFetchOptions',
   'runtimeOutputDirContext',
@@ -511,12 +503,7 @@ export const legacyFilenames = [
   'warningHandler',
   'windowsPath',
   'windowTitle',
-  'workspaceAgents',
   'workspaceContext',
-  'workspaceFileRead',
-  'workspaceFileSystem',
-  'workspaceFileWrite',
-  'workspaceMemory',
   'worktreeCleanup',
   'worktreeSessionService',
   'worktreeStartup',
@@ -525,10 +512,7 @@ export const legacyFilenames = [
   'CodeColorizer',
   'SKILL',
   'TeamManager',
-  'a2uiAction',
-  'acpSessionBridge',
   'asciiCharts',
-  'bridgeFileSystemAdapter',
   'btwUtils',
   'cdCommand',
   'claudeMcpImport',
@@ -537,12 +521,9 @@ export const legacyFilenames = [
   'compressFastCommand',
   'concurrencyLimiter',
   'configHash',
-  'connectionRegistry',
   'corruptFile',
   'cronTasksFile',
   'cronTasksLock',
-  'daemonLogger',
-  'daemonStatus',
   'enterPlanMode',
   'extensionPreferences',
   'forkCommand',
@@ -555,7 +536,6 @@ export const legacyFilenames = [
   'importConfigCommand',
   'inlineMediaLimit',
   'instructionsLoadedCallback',
-  'jsonRpc',
   'leaderPermissionBridge',
   'liveAgentPanelVisibility',
   'mcpApprovals',
@@ -564,7 +544,6 @@ export const legacyFilenames = [
   'mcpServers',
   'midTurnUserMessage',
   'normalizeDisabledTools',
-  'permissionAudit',
   'planApprovalGate',
   'promptAddendum',
   'retryErrorClassification',
@@ -578,7 +557,6 @@ export const legacyFilenames = [
   'settingsWatcher',
   'sleepInhibitor',
   'sourceRegistry',
-  'sseStream',
   'statsDataService',
   'tasksSnapshot',
   'teamHelpers',
@@ -587,15 +565,11 @@ export const legacyFilenames = [
   'toolCallIdUtils',
   'toolResultCleanup',
   'toolResultDisplayCompaction',
-  'transportStream',
   'usageHistoryService',
   'useMcpApproval',
   'useResizeSettleRepaint',
   'useStatsDialog',
   'useTeamInProcess',
   'userMemory',
-  'webShellStatic',
   'workflowsCommand',
-  'workspaceSettings',
-  'wsStream',
 ];
