+++ packages/cli/src/serve/acp-session-bridge.ts
@@ -10,10 +10,10 @@
  * #4175 PR F1 lifted the bridge core (`BridgeClient`,
  * `defaultSpawnChannelFactory`, `createAcpSessionBridge` factory closure,
  * plus the supporting types/errors/options/status) to
- * `@qwen-code/acp-bridge`. This shim preserves every existing relative
- * import path (`./acpSessionBridge.js`) so `server.ts`, `run-qwen-serve.ts`,
- * `workspace-agents.ts`, `workspace-memory.ts`, `index.ts`, plus the
- * bridge test suite, keep resolving without any call-site changes.
+ * `@qwen-code/acp-bridge`. This shim preserves the CLI-local bridge import
+ * surface so `server.ts`, `run-qwen-serve.ts`, `workspace-agents.ts`,
+ * `workspace-memory.ts`, `index.ts`, plus the bridge test suite, keep resolving
+ * through one module.
  *
  * The implementation now lives at:
  *   - `@qwen-code/acp-bridge/bridge` — `createAcpSessionBridge` factory
