+++ packages/cli/src/serve/daemon-logger.ts
