+++ packages/cli/src/serve/run-qwen-serve.ts
@@ -20,7 +20,7 @@ import {
   canonicalizeWorkspace,
   createAcpSessionBridge,
   type AcpSessionBridge,
-} from './acpSessionBridge.js';
+} from './acp-session-bridge.js';
 import {
   DEFAULT_OTLP_ENDPOINT,
   DEFAULT_TELEMETRY_TARGET,
@@ -51,7 +51,7 @@ import {
 import { createBridgeFileSystemAdapter } from './bridge-file-system-adapter.js';
 import { createDaemonStatusProvider } from './daemon-status-provider.js';
 import { isLoopbackBind } from './loopback-binds.js';
-import { resolveWebShellDir } from './webShellStatic.js';
+import { resolveWebShellDir } from './web-shell-static.js';
 import { parseAllowOriginPatterns } from './auth.js';
 import {
   createPermissionAuditPublisher,
@@ -62,7 +62,7 @@ import {
   getActiveSseCount,
   resolveBridgeFsFactory,
 } from './server.js';
-import { initDaemonLogger, type DaemonLogger } from './daemonLogger.js';
+import { initDaemonLogger, type DaemonLogger } from './daemon-logger.js';
 import { createSpawnChannelFactory } from '@qwen-code/acp-bridge/spawnChannel';
 import { createDaemonWorkspaceService } from './workspace-service/index.js';
 import { SERVE_CAPABILITY_REGISTRY } from './capabilities.js';
