+++ packages/cli/src/serve/web-shell-static.ts
