+++ packages/cli/src/serve/workspace-service/__tests__/integration.test.ts
@@ -20,7 +20,7 @@ import type {
   DaemonWorkspaceService,
   WorkspaceRequestContext,
 } from '../types.js';
-import type { AcpSessionBridge } from '../../acpSessionBridge.js';
+import type { AcpSessionBridge } from '../../acp-session-bridge.js';
 
 // ---------------------------------------------------------------------------
 // Test fixtures
